Rename JSONSourceSection to BundleSourceSection.
