+++ opengever/setup/cfgs/oggbundle.cfg
@@ -3,7 +3,7 @@ include =
     opengever.setup.content
 
 pipeline =
-    jsonsource
+    bundlesource
     disabled-initial-version
     resolveguid
     constructor
@@ -27,8 +27,8 @@ pipeline =
     savepoint
 
 
-[jsonsource]
-blueprint = opengever.setup.jsonsource
+[bundlesource]
+blueprint = opengever.setup.bundlesource
 
 [resolveguid]
 blueprint = opengever.setup.resolveguid
