+++ opengever/setup/tests/test_section_bundlesource.py
@@ -1,6 +1,6 @@
 from collective.transmogrifier.interfaces import ISection
 from collective.transmogrifier.interfaces import ISectionBlueprint
-from opengever.setup.sections.jsonsource import JSONSourceSection
+from opengever.setup.sections.bundlesource import BundleSourceSection
 from opengever.setup.tests import MockTransmogrifier
 from opengever.testing import FunctionalTestCase
 from pkg_resources import resource_filename
@@ -8,7 +8,7 @@ from zope.interface.verify import verifyClass
 from zope.interface.verify import verifyObject
 
 
-class TestJSONSource(FunctionalTestCase):
+class TestBundleSource(FunctionalTestCase):
 
     def setup_source(self, options, bundle_path=None, previous=None):
         previous = previous or []
@@ -17,13 +17,13 @@ class TestJSONSource(FunctionalTestCase):
         transmogrifier = MockTransmogrifier()
         transmogrifier.bundle_path = bundle_path
 
-        options.setdefault('blueprint', 'opengever.setup.jsonsource')
+        options.setdefault('blueprint', 'opengever.setup.bundlesource')
 
-        return JSONSourceSection(transmogrifier, '', options, previous)
+        return BundleSourceSection(transmogrifier, '', options, previous)
 
     def test_implements_interface(self):
-        self.assertTrue(ISection.implementedBy(JSONSourceSection))
-        verifyClass(ISection, JSONSourceSection)
+        self.assertTrue(ISection.implementedBy(BundleSourceSection))
+        verifyClass(ISection, BundleSourceSection)
 
-        self.assertTrue(ISectionBlueprint.providedBy(JSONSourceSection))
-        verifyObject(ISectionBlueprint, JSONSourceSection)
+        self.assertTrue(ISectionBlueprint.providedBy(BundleSourceSection))
+        verifyObject(ISectionBlueprint, BundleSourceSection)
